backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "GA4 OAuth AI Agent Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "openai-agents>=0.7.0",
    "supabase>=2.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.40.0",
    "google-auth-oauthlib>=1.2.0",
    "pydantic-settings>=2.0.0",
    "pyjwt[crypto]>=2.9.0",
]
